Adds ProgressBarActivity to avoid showProgress code repetition
