+++ app/src/main/java/utn/proy2k18/vantrack/initAndAccManagement/InitActivity.java
@@ -1,15 +1,10 @@
 package utn.proy2k18.vantrack.initAndAccManagement;
 
-import android.animation.Animator;
-import android.animation.AnimatorListenerAdapter;
-import android.annotation.TargetApi;
 import android.app.Activity;
 import android.content.Intent;
-import android.os.Build;
 import android.os.Bundle;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AlertDialog;
-import androidx.appcompat.app.AppCompatActivity;
 import android.util.Log;
 import android.view.View;
 import android.widget.Button;
@@ -37,7 +32,7 @@ import utn.proy2k18.vantrack.exceptions.FailedToCreateUserException;
 import utn.proy2k18.vantrack.models.User;
 import utn.proy2k18.vantrack.viewModels.UsersViewModel;
 
-public class InitActivity extends AppCompatActivity {
+public class InitActivity extends ProgressBarActivity {
 
     private static final String TAG = "GoogleActivity";
     private static final String ARG_PARAM1 = "avoid_automatic_login";
@@ -51,8 +46,6 @@ public class InitActivity extends AppCompatActivity {
     private Boolean avoidAutomaticLogin = false;
     private Boolean signOut = false;
     private Boolean revokeAccess = false;
-    private View mProgressView;
-    private View mInitFormView;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -239,47 +232,11 @@ public class InitActivity extends AppCompatActivity {
             }
         });
 
-        mInitFormView = findViewById(R.id.init_form);
+        mFormView = findViewById(R.id.init_form);
         mProgressView = findViewById(R.id.init_progress);
     }
 
     private void goToActivity(Class<?> activityToOpen) {
         startActivity(new Intent(this, activityToOpen));
-    }
-
-    /**
-     * Shows the progress UI and hides the login form.
-     */
-    @TargetApi(Build.VERSION_CODES.HONEYCOMB_MR2)
-    private void showProgress(final boolean show) {
-        // On Honeycomb MR2 we have the ViewPropertyAnimator APIs, which allow
-        // for very easy animations. If available, use these APIs to fade-in
-        // the progress spinner.
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.HONEYCOMB_MR2) {
-            int shortAnimTime = getResources().getInteger(android.R.integer.config_shortAnimTime);
-
-            mInitFormView.setVisibility(show ? View.GONE : View.VISIBLE);
-            mInitFormView.animate().setDuration(shortAnimTime).alpha(
-                    show ? 0 : 1).setListener(new AnimatorListenerAdapter() {
-                @Override
-                public void onAnimationEnd(Animator animation) {
-                    mInitFormView.setVisibility(show ? View.GONE : View.VISIBLE);
-                }
-            });
-
-            mProgressView.setVisibility(show ? View.VISIBLE : View.GONE);
-            mProgressView.animate().setDuration(shortAnimTime).alpha(
-                    show ? 1 : 0).setListener(new AnimatorListenerAdapter() {
-                @Override
-                public void onAnimationEnd(Animator animation) {
-                    mProgressView.setVisibility(show ? View.VISIBLE : View.GONE);
-                }
-            });
-        } else {
-            // The ViewPropertyAnimator APIs are not available, so simply show
-            // and hide the relevant UI components.
-            mProgressView.setVisibility(show ? View.VISIBLE : View.GONE);
-            mInitFormView.setVisibility(show ? View.GONE : View.VISIBLE);
-        }
     }
 }
