+++ app/src/main/java/utn/proy2k18/vantrack/initAndAccManagement/SignUpActivity.java
@@ -1,14 +1,9 @@
 package utn.proy2k18.vantrack.initAndAccManagement;
 
-import android.animation.Animator;
-import android.animation.AnimatorListenerAdapter;
-import android.annotation.TargetApi;
 import android.app.Activity;
-import android.os.Build;
 import android.os.Bundle;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AlertDialog;
-import androidx.appcompat.app.AppCompatActivity;
 import android.util.Log;
 import android.view.View;
 import android.widget.Button;
@@ -25,7 +20,7 @@ import utn.proy2k18.vantrack.exceptions.FailedToCreateUserException;
 import utn.proy2k18.vantrack.models.User;
 import utn.proy2k18.vantrack.viewModels.UsersViewModel;
 
-public class SignUpActivity extends AppCompatActivity {
+public class SignUpActivity extends ProgressBarActivity {
 
     private FirebaseAuth mAuth;
     private TextView name;
@@ -36,8 +31,6 @@ public class SignUpActivity extends AppCompatActivity {
     private TextView password;
     private TextView passwordCopy;
     private String errorMsg;
-    private View mProgressView;
-    private View mSignUpFormView;
     private UsersViewModel usersViewModel;
 
     @Override
@@ -143,7 +136,7 @@ public class SignUpActivity extends AppCompatActivity {
             }
         });
 
-        mSignUpFormView = findViewById(R.id.sign_up_form);
+        mFormView = findViewById(R.id.sign_up_form);
         mProgressView = findViewById(R.id.sign_up_progress);
     }
 
@@ -223,41 +216,5 @@ public class SignUpActivity extends AppCompatActivity {
                 .setNeutralButton("Aceptar",null)
                 .create();
         alertDialog.show();
-    }
-
-    /**
-     * Shows the progress UI and hides the login form.
-     */
-    @TargetApi(Build.VERSION_CODES.HONEYCOMB_MR2)
-    private void showProgress(final boolean show) {
-        // On Honeycomb MR2 we have the ViewPropertyAnimator APIs, which allow
-        // for very easy animations. If available, use these APIs to fade-in
-        // the progress spinner.
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.HONEYCOMB_MR2) {
-            int shortAnimTime = getResources().getInteger(android.R.integer.config_shortAnimTime);
-
-            mSignUpFormView.setVisibility(show ? View.GONE : View.VISIBLE);
-            mSignUpFormView.animate().setDuration(shortAnimTime).alpha(
-                    show ? 0 : 1).setListener(new AnimatorListenerAdapter() {
-                @Override
-                public void onAnimationEnd(Animator animation) {
-                    mSignUpFormView.setVisibility(show ? View.GONE : View.VISIBLE);
-                }
-            });
-
-            mProgressView.setVisibility(show ? View.VISIBLE : View.GONE);
-            mProgressView.animate().setDuration(shortAnimTime).alpha(
-                    show ? 1 : 0).setListener(new AnimatorListenerAdapter() {
-                @Override
-                public void onAnimationEnd(Animator animation) {
-                    mProgressView.setVisibility(show ? View.VISIBLE : View.GONE);
-                }
-            });
-        } else {
-            // The ViewPropertyAnimator APIs are not available, so simply show
-            // and hide the relevant UI components.
-            mProgressView.setVisibility(show ? View.VISIBLE : View.GONE);
-            mSignUpFormView.setVisibility(show ? View.GONE : View.VISIBLE);
-        }
     }
 }
